+++ resources/views/flights/pastFlights.blade.php
@@ -60,7 +60,7 @@
                                 <a href="{{ route('editFlightForm', $pastFlight->id) }}" class="crudBtn">
                                     <img src="{{asset('img/edit.png') }}" alt="edit-Button" class="crudBtn">
                                 </a>
-                                <a href="?action=delete&id={{ $pastFlight->id }}">
+                                <a href="?action=delete&id={{ $pastFlight->id }}"onclick="return confirm('Are you sure you want to delete this flight? ID: {{ $pastFlight->id }}')">
                                     <img src="{{ asset('img/delete.png') }}" alt="delete-Button" class="crudBtn">
                                 </a>
                             </td>
